add suspense to dashboard page
