+++ lib/actions/audits.ts
@@ -1,5 +1,11 @@
 import { prisma } from "@/lib/db/prisma.server";
 import { Audit, User, Profile, Terms, PrismaPromise } from "@prisma/client";
+import fs from "fs";
+import path from "path";
+import { remark } from "remark";
+import html from "remark-html";
+import remarkGfm from "remark-gfm";
+import matter from "gray-matter";
 
 interface AuditsRelation extends Audit {
   auditee: User & {
@@ -93,4 +99,24 @@ export const getAudit = (id: string): PrismaPromise<AuditRelation> => {
       },
     },
   });
+};
+
+export const getMarkdown = async (display: string): Promise<string> => {
+  if (!["details", "audit"].includes(display)) {
+    display = "details";
+  }
+
+  let filePath: string;
+  console.log(path.resolve("./public", display + ".md"));
+  if (process.env.NODE_ENV === "development") {
+    filePath = path.resolve("public", display + ".md");
+  } else {
+    filePath = path.resolve("./public", display + ".md");
+  }
+  const fileContents = fs.readFileSync(filePath, "utf8");
+
+  const { content } = matter(fileContents);
+  const processedContent = (await remark().use(html).use(remarkGfm).process(content)).toString();
+
+  return processedContent;
 };
